move wait_for_finish
